Further refactoring of DWI::Directions
- Define default FOD FMLS segmenter direction set once, and refer to it in all locations where FOD segmentation may occur.
- Define new class DWI::Directions::Adjacency, which just encapsulates the adjacency of a direction set and does not store the set itself.
- Define new class DWI::Directions::CartesianWithAdjacency; this provides the same functionality as the old DWI::Directions::Set, with the unused masking members left out, and is more explicit about what is contained.
