+++ src/dwi/tractography/mapping/mapper_plugins.h
@@ -39,7 +39,7 @@ namespace MR {
 
 
         class DixelMappingPlugin
-        { 
+        {
           public:
             DixelMappingPlugin (const DWI::Directions::FastLookupSet& directions) :
               dirs (directions) { }
@@ -53,7 +53,7 @@ namespace MR {
 
 
         class TODMappingPlugin
-        { 
+        {
           public:
             TODMappingPlugin (const size_t N) :
               generator (new Math::SH::aPSF<float> (Math::SH::LforN (N))) { }
@@ -68,7 +68,7 @@ namespace MR {
 
 
         class TWIImagePluginBase
-        { 
+        {
           public:
             TWIImagePluginBase (const std::string& input_image, const tck_stat_t track_statistic) :
                 statistic (track_statistic),
@@ -117,7 +117,7 @@ namespace MR {
 
 
         class TWIScalarImagePlugin : public TWIImagePluginBase
-        { 
+        {
           public:
             TWIScalarImagePlugin (const std::string& input_image, const tck_stat_t track_statistic) :
                 TWIImagePluginBase (input_image, track_statistic)
@@ -148,7 +148,7 @@ namespace MR {
 
 
         class TWIFODImagePlugin : public TWIImagePluginBase
-        { 
+        {
           public:
             TWIFODImagePlugin (const std::string& input_image, const tck_stat_t track_statistic) :
                 TWIImagePluginBase (input_image, track_statistic),
@@ -179,7 +179,7 @@ namespace MR {
 
 
         class TWDFCStaticImagePlugin : public TWIImagePluginBase
-        { 
+        {
           public:
             TWDFCStaticImagePlugin (Image<float>& input_image) :
                 TWIImagePluginBase (input_image, ENDS_CORR) { }
@@ -198,7 +198,7 @@ namespace MR {
 
 
         class TWDFCDynamicImagePlugin : public TWIImagePluginBase
-        { 
+        {
           public:
             TWDFCDynamicImagePlugin (Image<float>& input_image, const vector<float>& kernel, const ssize_t timepoint) :
                 TWIImagePluginBase (input_image, ENDS_CORR),
