+++ src/dwi/tractography/seeding/dynamic.cpp
@@ -65,7 +65,7 @@ namespace MR
 
 
 
-      Dynamic::Dynamic (const std::string& in, Image<float>& fod_data, const size_t num, const DWI::Directions::FastLookupSet& dirs) :
+      Dynamic::Dynamic (const std::string& in, Image<float>& fod_data, const size_t num, const DWI::Directions::CartesianWithAdjacency& dirs) :
           Base (in, "dynamic", MAX_TRACKING_SEED_ATTEMPTS_DYNAMIC),
           SIFT::ModelBase<Fixel_TD_seed> (fod_data, dirs),
           target_trackcount (num),
