+++ src/gui/mrview/tool/odf/item.cpp
@@ -19,6 +19,7 @@
 #include "header.h"
 #include "dwi/gradient.h"
 #include "math/SH.h"
+#include "dwi/directions/predefined.h"
 
 namespace MR
 {
@@ -85,7 +86,7 @@ namespace MR
           auto entry = H.keyval().find ("directions");
           if (entry != H.keyval().end()) {
             try {
-              header_dirs = deserialize_matrix<float> (entry->second);
+              header_dirs = deserialize_matrix (entry->second);
               Math::Sphere::check (header_dirs, H.size (3));
             } catch (Exception& e) {
               DEBUG (e[0]);
@@ -99,11 +100,11 @@ namespace MR
             throw Exception ("No valid DW scheme defined in header");
           if (index >= shells->count())
             throw Exception ("Shell index is outside valid range");
-          Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic> shell_dirs ((*shells)[index].count(), 3);
+          Eigen::Matrix<default_type, Eigen::Dynamic, Eigen::Dynamic> shell_dirs ((*shells)[index].count(), 3);
           const vector<size_t>& volumes = (*shells)[index].get_volumes();
           for (size_t row = 0; row != volumes.size(); ++row)
-            shell_dirs.row (row) = grad.row (volumes[row]).head<3>().cast<float>();
-          auto new_dirs = MR::make_unique<MR::DWI::Directions::Set> (shell_dirs);
+            shell_dirs.row (row) = grad.row (volumes[row]).head<3>();
+          auto new_dirs = MR::make_unique<MR::DWI::Directions::CartesianWithAdjacency> (shell_dirs);
           std::swap (dirs, new_dirs);
           shell_index = index;
           dir_type = DixelPlugin::dir_t::DW_SCHEME;
@@ -112,13 +113,13 @@ namespace MR
         void ODF_Item::DixelPlugin::set_header() {
           if (!header_dirs.rows())
             throw Exception ("No direction scheme defined in header");
-          auto new_dirs = MR::make_unique<MR::DWI::Directions::Set> (header_dirs);
+          auto new_dirs = MR::make_unique<MR::DWI::Directions::CartesianWithAdjacency> (header_dirs);
           std::swap (dirs, new_dirs);
           dir_type = DixelPlugin::dir_t::HEADER;
         }
 
         void ODF_Item::DixelPlugin::set_internal (const size_t n) {
-          auto new_dirs = MR::make_unique<MR::DWI::Directions::Set> (n);
+          auto new_dirs = MR::make_unique<MR::DWI::Directions::CartesianWithAdjacency> (MR::DWI::Directions::load (n));
           std::swap (dirs, new_dirs);
           dir_type = DixelPlugin::dir_t::INTERNAL;
         }
@@ -131,7 +132,7 @@ namespace MR
 
         void ODF_Item::DixelPlugin::set_from_file (const std::string& path)
         {
-          auto new_dirs = MR::make_unique<MR::DWI::Directions::Set> (path);
+          auto new_dirs = MR::make_unique<MR::DWI::Directions::CartesianWithAdjacency> (MR::DWI::Directions::load (path, true));
           std::swap (dirs, new_dirs);
           dir_type = DixelPlugin::dir_t::FILE;
         }
